+++ asdf/_tests/commands/tests/test_defragment.py
@@ -53,3 +53,8 @@ def test_defragment_bzp2(tmpdir):
 def test_defragment_lz4(tmpdir):
     pytest.importorskip("lz4")
     _test_defragment(tmpdir, "lz4")
+
+
+def test_defragment_blosc(tmpdir):
+    pytest.importorskip("blosc")
+    _test_defragment(tmpdir, "blosc")