+++ asdf/_tests/test_array_blocks.py
@@ -858,7 +858,7 @@ def test_add_block_before_fully_loaded(tmp_path):
 
 
 @pytest.mark.parametrize("all_array_storage", ["internal", "external", "inline"])
-@pytest.mark.parametrize("all_array_compression", [None, "", "zlib", "bzp2", "lz4", "input"])
+@pytest.mark.parametrize("all_array_compression", [None, "", "zlib", "bzp2", "lz4", "blosc", "input"])
 @pytest.mark.parametrize("compression_kwargs", [None, {}])
 def test_write_to_update_storage_options(tmp_path, all_array_storage, all_array_compression, compression_kwargs):
     if all_array_compression == "bzp2" and compression_kwargs is not None: