+++ asdf/asdf.py
@@ -674,6 +674,8 @@ class AsdfFile:
 
             - ``lz4``: Use lz4 compression
 
+            - ``blosc``: Use lz4 compression
+
             - ``input``: Use the same compression as in the file read.
               If there is no prior file, acts as None.
 
@@ -1003,6 +1005,8 @@ class AsdfFile:
 
             - ``lz4``: Use lz4 compression.
 
+            - ``blosc``: Use blosc compression.
+
             - ``input``: Use the same compression as in the file read.
               If there is no prior file, acts as None
 
@@ -1155,6 +1159,8 @@ class AsdfFile:
             - ``bzp2``: Use bzip2 compression.
 
             - ``lz4``: Use lz4 compression.
+
+            - ``blosc``: Use blosc compression.
 
             - ``input``: Use the same compression as in the file read.
               If there is no prior file, acts as None.