+++ asdf/extension/_serialization_context.py
@@ -203,6 +203,8 @@ class SerializationContext:
 
             - ``lz4``: Use lz4 compression
 
+            - ``blosc``: Use blosc compression
+
             - ``input``: Use the same compression as in the file read.
               If there is no prior file, acts as None.